requirements.txt
astroid==2.0.4
certifi==2018.10.15
chardet==3.0.4
cloudinary==1.14.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
Django==2.1.3
django-allauth==0.38.0
django-cors-headers==2.4.0
django-extra-fields==1.0.0
django-rest-auth==0.9.3
djangorestframework==3.9.0
drf-nested-routers==0.91
idna==2.7
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
Markdown==3.0.1
MarkupSafe==1.0
mccabe==0.6.1
mock==2.0.0
numpy==1.15.3
oauthlib==2.1.0
pbr==5.1.0
Pillow==5.3.0
Pygments==2.2.0
pylint==2.1.1
pylint-django==2.0.2
pylint-plugin-utils==0.4
python-dateutil==2.7.5
python3-openid==3.1.0
pytz==2018.7
requests==2.20.0
requests-oauthlib==1.0.0
six==1.11.0
typed-ast==1.1.0
uritemplate==3.0.0
urllib3==1.24
wrapt==1.10.11
